requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
astunparse==1.6.3
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.64.1
gTTS==2.5.1
h5py==3.11.0
idna==3.7
importlib_metadata==8.0.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.4.1
keras-utils==1.0.13
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
murmurhash==1.0.10
namex==0.0.8
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pathlib==1.0.1
pillow==10.4.0
playsound==1.3.0
preshed==3.0.9
protobuf==4.25.3
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.0
Pygments==2.18.0
pyserial==3.5
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.1
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thinc==8.2.5
threadpoolctl==3.5.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
